agent/feishu-agent/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
openai>=1.12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
pycryptodome>=3.20.0
watchfiles>=0.21.0
pytest>=7.4.0
lark-oapi>=1.5.0
